Cleanup
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@6973 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/Extensions/org/objectweb/proactive/extensions/scilab/test/SciTest4.java
@@ -88,8 +88,8 @@ public class SciTest4 {
             m1 = new double[nbRow * nbCol];
             m2 = new double[nbRow * nbCol];
             for (int i = 0; i < (nbRow * nbCol); i++) {
-                m1[i] = (double) (Math.random() * 10);
-                m2[i] = (double) (Math.random() * 10);
+                m1[i] = (Math.random() * 10);
+                m2[i] = (Math.random() * 10);
             }
 
             SciDoubleMatrix sciMatrix1 = new SciDoubleMatrix("M1", nbRow,
